版本 1.1.3.17 更新及功能增强
更新了 `CHANGELOG.md` 文件，增加了版本 `1.1.3.17` 的更新日志，描述了修复安装包缺失文件的问题，并在菜单中增加了重新安装当前软件的选项。
在 `MenuConfigExport.cs` 文件中，修改了注释，明确了获取当前安装文件路径的操作。
在 `AutoUpdater.cs` 文件中，增加了对特定版本更新文件的处理逻辑，将其移动到应用程序数据文件夹中，并删除了不再需要的更新文件。同时，将 `RestartApplication` 方法的访问修饰符从 `private` 改为 `public`。
在 `ChangelogWindow.xaml.cs` 文件中，移除了不再使用的 `using` 语句和一些菜单项类。
删除了 `UpdateJob.cs` 文件中的 `UpdateJob` 类及其相关逻辑。
在 `WindowUpdate.xaml.cs` 文件中，增加了 `WindowUpdate` 类的部分注释。
在 `Directory.Build.props` 文件中，将版本前缀从 `1.1.3.16` 更新为 `1.1.3.17`。
在 `ExportEventWindow.cs` 文件中，移除了 `ExportUploadWindow` 类及其相关逻辑。
新增了 `MenuChangeLog.cs` 文件，定义了 `MenuChangeLog` 类，用于显示更新日志窗口。
新增了 `MenuForceUpdate.cs` 文件，定义了 `MenuForceUpdate` 类，用于强制更新操作。
新增了 `MenuUpdate.cs` 文件，定义了 `MenuUpdate` 类，用于更新操作。
新增了 `MenuUpdateBaidu.cs` 文件，定义了 `MenuUpdateBaidu` 类，用于通过百度云下载更新。
新增了 `MenuUpdatelocal.cs` 文件，定义了 `MenuUpdatelocal` 类，用于本地下载更新。
新增了 `MenuUpdateReInstallPath.cs` 文件，定义了 `MenuUpdateReInstallPath` 类，用于打开缓存文件夹。
新增了 `MenuUpload.cs` 文件，定义了 `MenuUpload` 类，用于上传文件操作。

+++ ColorVision/Update/AutoUpdater.cs
@@ -92,6 +92,14 @@ namespace ColorVision.Update
 
             foreach (string updateFile in updateFiles)
             {
+                if (updateFile.Contains($"ColorVision-{CurrentVersion}.exe"))
+                {
+                    string AssemblyCompany = Assembly.GetExecutingAssembly().GetCustomAttribute<AssemblyCompanyAttribute>()?.Company ?? "ColorVision";
+
+                    File.Move(updateFile, Path.Combine(Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData),AssemblyCompany), $"ColorVision-{CurrentVersion}.exe"));
+                    continue;
+                }
+
                 try
                 {
                     File.Delete(updateFile);
@@ -373,7 +381,7 @@ namespace ColorVision.Update
         }
 
 
-        private static void RestartApplication(string downloadPath)
+        public static void RestartApplication(string downloadPath)
         {
             // 保存数据库配置
             ConfigHandler.GetInstance().SaveConfigs();
